refactor(allowed_modules): use the new modules API
