+++ frappe/boot.py
@@ -96,8 +96,8 @@ def load_conf_settings(bootinfo):
 		if key in conf: bootinfo[key] = conf.get(key)
 
 def load_desktop_icons(bootinfo):
-	from frappe.desk.doctype.desktop_icon.desktop_icon import get_modules_from_all_apps
-	bootinfo.allowed_modules = get_modules_from_all_apps()
+	from frappe.config import get_modules_from_all_apps_for_user
+	bootinfo.allowed_modules = get_modules_from_all_apps_for_user()
 
 def get_allowed_pages():
 	return get_user_pages_or_reports('Page')
