+++ frappe/core/doctype/user/user.json
@@ -1511,7 +1511,7 @@
    "columns": 0, 
    "default": "", 
    "description": "", 
-   "fieldname": "desktop_icon_access", 
+   "fieldname": "sb_allow_modules", 
    "fieldtype": "Section Break", 
    "hidden": 0, 
    "ignore_user_permissions": 0, 
@@ -1520,7 +1520,7 @@
    "in_global_search": 0, 
    "in_list_view": 0, 
    "in_standard_filter": 0, 
-   "label": "Allow Desktop Icon", 
+   "label": "Allow Modules", 
    "length": 0, 
    "no_copy": 0, 
    "permlevel": 1, 
@@ -2303,7 +2303,7 @@
  "issingle": 0, 
  "istable": 0, 
  "max_attachments": 5, 
- "modified": "2018-11-21 12:34:57.652854", 
+ "modified": "2019-01-30 13:56:10.732154", 
  "modified_by": "Administrator", 
  "module": "Core", 
  "name": "User",
