+++ frappe/desk/doctype/desktop_icon/desktop_icon.py
@@ -487,44 +487,3 @@ def hide(name, user = None):
 		return False
 
 	return True
-
-@frappe.whitelist()
-def get_modules_from_all_apps():
-	modules_list = []
-	for app in frappe.get_installed_apps():
-		modules_list += get_modules_from_app(app)
-	return modules_list
-
-def get_modules_from_app(app):
-	try:
-		modules = frappe.get_attr(app + '.config.desktop.get_data')() or {}
-	except ImportError:
-		return []
-
-	# Only newly formatted modules that have a category to be shown on desk
-	modules = [m for m in modules if m.get("category")]
-
-	active_domains = frappe.get_active_domains()
-
-	if isinstance(modules, dict):
-		allowed_modules_list = []
-		for m, desktop_icon in iteritems(modules):
-			desktop_icon['module_name'] = m
-			allowed_modules_list.append(desktop_icon)
-	else:
-		allowed_modules_list = []
-		for m in modules:
-			to_add = True
-
-			# Check Domain
-			if is_domain(m):
-				if m.get("module_name") not in active_domains:
-					to_add = False
-
-			if to_add:
-				allowed_modules_list.append(m)
-
-	return allowed_modules_list
-
-def is_domain(module):
-	return module.get("category") == "Domains"
